Extract Client Gateway cache invalidation
